requirements.txt
python-telegram-bot[job-queue]
Flask
pymongo
requests
PyGithub
